Visual - Introduce VEntityEnemyCommon.state and EnemyState enum, improve neighbor filtering in astar_search(), replace for..in loop with if len(collided) > 0 in VGame.resolve_player_enemy_collisions()

+++ src/visual/vgame.py
@@ -266,7 +266,7 @@ class VGame(arcade.View):
         collided: list[VEntityEnemyCommon] = (
             arcade.check_for_collision_with_list(self.player, self.enemy_list)
         )
-        for _enemy in collided:
+        if len(collided) > 0:
             self.player.kill()
 
     # ########################################################################
